Duplicating the work of PR #389 https://github.com/stephencelis/SQLite.swift/pull/389
Needed to make func check() be public, so that SQLiteCipher.swift can access it in it's extension of the Connection class.

Please let me know what questions you have. if there is a better way of doing this, I'm open to suggestions. If you want to check out the details of the PR I have coming for the SQLiteCipher.swift repository, that uses this change, you can check it out here: https://github.com/gblotter/SQLite.swift/tree/SQLiteCipher-play-nice